# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: string;
  name: string;
  type: string;
  startsAt: string;
  location: string;
  description: string;
  capacity: number;
  coverUrl: string;
  inviteCount: number;
  checkedInCount: number;
  status: string;
}

export interface EventInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  startsAt: string;
  location: string;
  description: string;
  /** @minimum 1 */
  capacity: number;
  coverUrl: string;
}

export interface Invite {
  id: string;
  eventId: string;
  guestName: string;
  /** @nullable */
  contact: string | null;
  code: string;
  status: string;
  /** @nullable */
  checkedInAt: string | null;
  createdAt: string;
}

export type EventDetail = Event & {
  invites: Invite[];
};

export interface InviteInput {
  /** @minLength 1 */
  guestName: string;
  /** @nullable */
  contact?: string | null;
}

export interface ValidationResult {
  valid: boolean;
  message: string;
  invite: Invite;
}