# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Event management
  - name: invites
    description: Invitation management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List organizer events
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event with invitations
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Event not found
  /events/{id}/invites:
    post:
      operationId: createInvite
      tags: [invites]
      summary: Add a guest and create a unique invitation
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Created invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "409":
          description: Event capacity reached
  /invites/{code}/cancel:
    post:
      operationId: cancelInvite
      tags: [invites]
      summary: Cancel an invitation
      parameters:
        - $ref: "#/components/parameters/InviteCode"
      responses:
        "200":
          description: Cancelled invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /invites/{code}/validate:
    post:
      operationId: validateInvite
      tags: [invites]
      summary: Validate and check in an invitation
      parameters:
        - $ref: "#/components/parameters/InviteCode"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
components:
  parameters:
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
    InviteCode:
      name: code
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Event:
      type: object
      required: [id, name, type, startsAt, location, description, capacity, coverUrl, inviteCount, checkedInCount, status]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        startsAt: { type: string, format: date-time }
        location: { type: string }
        description: { type: string }
        capacity: { type: number }
        coverUrl: { type: string }
        inviteCount: { type: number }
        checkedInCount: { type: number }
        status: { type: string }
    EventInput:
      type: object
      required: [name, type, startsAt, location, description, capacity, coverUrl]
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string, minLength: 1 }
        startsAt: { type: string, format: date-time }
        location: { type: string }
        description: { type: string }
        capacity: { type: number, minimum: 1 }
        coverUrl: { type: string }
    Invite:
      type: object
      required: [id, eventId, guestName, contact, code, status, checkedInAt, createdAt]
      properties:
        id: { type: string }
        eventId: { type: string }
        guestName: { type: string }
        contact: { type: ["string", "null"] }
        code: { type: string }
        status: { type: string }
        checkedInAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          required: [invites]
          properties:
            invites:
              type: array
              items:
                $ref: "#/components/schemas/Invite"
    InviteInput:
      type: object
      required: [guestName]
      properties:
        guestName: { type: string, minLength: 1 }
        contact: { type: ["string", "null"] }
    ValidationResult:
      type: object
      required: [valid, message, invite]
      properties:
        valid: { type: boolean }
        message: { type: string }
        invite:
          $ref: "#/components/schemas/Invite"