# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: string;
  name: string;
  type: string;
  startsAt: Date;
  location: string;
  description: string;
  capacity: number;
  coverUrl: string;
  inviteCount: number;
  checkedInCount: number;
  status: string;
}