# lib/api-zod/src/generated/types/eventDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from './event';
import type { Invite } from './invite';

export type EventDetail = Event & {
  invites: Invite[];
};