# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  startsAt: Date;
  location: string;
  description: string;
  /** @minimum 1 */
  capacity: number;
  coverUrl: string;
}