# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Invite {
  id: string;
  eventId: string;
  guestName: string;
  /** @nullable */
  contact: string | null;
  code: string;
  status: string;
  /** @nullable */
  checkedInAt: Date | null;
  createdAt: Date;
}