# lib/api-zod/src/generated/types/validationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Invite } from './invite';

export interface ValidationResult {
  valid: boolean;
  message: string;
  invite: Invite;
}